backend/microservices/auth_services/pyproject.toml
[tool.poetry]
name = "auth-services"
version = "0.1.0"
description = "One of the microservice, containg the authentication logic of the website/application"
authors = ["eedu7 <mueedahmad067@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.8"}
sqlalchemy = "^2.0.37"
pydantic = "^2.10.6"
pydantic-settings = "^2.7.1"
asyncpg = "^0.30.0"
python-jose = "^3.3.0"
alembic = "^1.14.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-extra-types = "^2.10.2"
phonenumbers = "^8.13.54"
boto3 = "^1.36.14"
redis = "^5.2.1"
celery = "^5.4.0"
faker = "^36.1.0"
pytest-asyncio = "^0.25.3"
pylint = "^3.3.4"


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.4"
mypy = "^1.14.1"
icecream = "^2.1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]  # F for unused imports, I for sorting
ignore = ["E501"]  # Ignore long lines

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
preview = true
